requirements.txt
mne>=1.4.0
numpy>=1.20.2
matplotlib==3.7.2
setuptools==65.5.1
scipy>=1.10.1
inflection~=0.5.1
pytest~=6.2.5
pandas~=2.0.3
pymatreader~=0.0.32
boto3~=1.28.43
botocore~=1.31.43
torch~=2.0.1
geovoronoi~=0.4.0
shapely~=2.0.1
enlighten~=1.11.2
scikit-learn~=1.3.0
PyYAML~=6.0.1
Braindecode~=0.8
seaborn~=0.13.2
umap-learn~=0.5.5